refactor: migración a Supabase Cloud - eliminación de configuración local
- ✅ Eliminada toda la configuración de Supabase local (backend/supabase/)
- ✅ Eliminados archivos de configuración local (mcp-server-local, scripts locales)
- ✅ Preservadas migraciones SQL en docs/base_datos/migraciones/
- ✅ Mantenidos todos los modelos de datos (frontend/lib/models/)
- ✅ Mantenidos todos los servicios (frontend/lib/services/)
- ✅ Mantenidos todos los BLoCs (frontend/lib/blocs/)
- ✅ Actualizado README.md para reflejar uso de Supabase Cloud
- ✅ Marcado docs/desarrollo-local.md como obsoleto
- ✅ Creada rama de backup: backup-supabase-local

BREAKING CHANGE: El proyecto ahora requiere Supabase Cloud.
Para configuración local, usar rama: backup-supabase-local

+++ frontend/lib/models/task.dart
@@ -19,7 +19,7 @@ class Task {
   @JsonKey(name: 'completed_at')
   final DateTime? completedAt;
   @JsonKey(name: 'kanban_position')
-  final int? kanbanPosition;
+  final double? kanbanPosition;
   @JsonKey(name: 'estimated_hours')
   final int? estimatedHours;
   @JsonKey(name: 'actual_hours')
@@ -68,7 +68,7 @@ class Task {
     TaskStatus? status,
     DateTime? dueDate,
     DateTime? completedAt,
-    int? kanbanPosition,
+    double? kanbanPosition,
     int? estimatedHours,
     int? actualHours,
     TaskComplexity? complexity,
